testing: use noplogger
